replace unused dropdown component with new one
- remove che-button-dropdown
- rename che-button-dropdown2 => che-button-dropdown
