+++ dashboard/src/app/demo-components/demo-components.html
@@ -102,24 +102,15 @@ Ultima Syriarum est Palaestina per intervalla magna protenta, cultis abundans te
       <div layout="column"
            style="padding-bottom: 100px;">
 
-        <!-- Old dropdown button -->
-        <textarea demo-source-render>
-          <che-button-dropdown che-button-dropdown-label="Dropdown button"
-                               che-button-dropdown-href="{{demoComponentsController.dropdownHref}}"
-                               che-button-dropdown-controller="demoComponentsController"
-                               che-disabled="false"></che-button-dropdown>
-        </textarea>
-
-        <!-- New dropdown button -->
-        <label for="dropdown2disabled">The dropdown button is disabled:
-          <input id="dropdown2disabled" type="checkbox" ng-model="dropdown2disabled">
+        <label for="dropdownDisabled">The dropdown button is disabled:
+          <input id="dropdownDisabled" type="checkbox" ng-model="dropdownDisabled">
         </label>
 
         <textarea demo-source-render>
-          <che-button-dropdown2 button-style="che-button-primary"
-                                main-action-config="demoComponentsController.buttonDropdownConfig.mainAction"
-                                other-actions-config="demoComponentsController.buttonDropdownConfig.otherActions"
-                                button-disabled="dropdown2disabled"></che-button-dropdown2>
+          <che-button-dropdown button-style="che-button-primary"
+                               main-action-config="demoComponentsController.buttonDropdownConfig.mainAction"
+                               other-actions-config="demoComponentsController.buttonDropdownConfig.otherActions"
+                               button-disabled="dropdownDisabled"></che-button-dropdown>
         </textarea>
 
       </div>
