+++ dashboard/src/app/workspaces/create-workspace/create-workspace.html
@@ -1,9 +1,9 @@
 <che-toolbar che-title="New Workspace">
-  <che-button-dropdown2 button-style="che-button-save-flat"
-                        button-disabled="createWorkspaceController.isCreateButtonDisabled()"
-                        main-action-config="createWorkspaceController.headerCreateButtonConfig.mainAction"
-                        other-actions-config="createWorkspaceController.headerCreateButtonConfig.otherActions"
-                        class="create-workspace-header-button"></che-button-dropdown2>
+  <che-button-dropdown button-style="che-button-save-flat"
+                       button-disabled="createWorkspaceController.isCreateButtonDisabled()"
+                       main-action-config="createWorkspaceController.headerCreateButtonConfig.mainAction"
+                       other-actions-config="createWorkspaceController.headerCreateButtonConfig.otherActions"
+                       class="create-workspace-header-button"></che-button-dropdown>
 </che-toolbar>
 
 <md-progress-linear md-mode="indeterminate"
