+++ dashboard/src/app/workspaces/create-workspace/create-workspace.styl
@@ -3,7 +3,7 @@
   top 56px
   z-index 2
 
-.create-workspace-header-button .che-button-dropdown2
+.create-workspace-header-button .che-button-dropdown
   button.che-button,
   .fa-caret-down
     min-height 36px
