+++ dashboard/src/components/widget/button-dropdown/che-button-dropdown.directive.ts
@@ -10,24 +10,110 @@
  */
 'use strict';
 
-/**
- * Defines the super class for for all buttons
- * @author Florent Benoit
- */
-export class CheButtonDropdown implements ng.IDirective {
+/* tslint:disable */
+type ButtonType = 'link' | 'button';
+/* tslint:enable */
 
+export interface ICheButtonDropdownMainAction {
+  action?: (args?: any) => void;
+  disabled?: boolean;
+  href?: string;
+  id?: string;
+  name?: string;
+  title: string;
+  type: ButtonType;
+}
+export interface ICheButtonDropdownOtherAction extends ICheButtonDropdownMainAction {
+  orderNumber: number;
+}
+
+interface ICheButtonDropdownAttributes extends ng.IAttributes {
+  mainActionConfig: string;
+  otherActionsConfig: string;
+  buttonDisabled: string;
+  buttonStyle: string;
+}
+
+/**
+ * @ngdoc directive
+ * @name components.directive:cheButtonDropdown
+ * @restrict E
+ *
+ * @description
+ * `<che-button-dropdown>` defines a dropdown button
+ *
+ * @param {ICheButtonDropdownMainAction} main-action-config configuration of the visible button (split-button)
+ * @param {Array<ICheButtonDropdownOtherAction>} other-actions-config configuration of the dropdown menu items
+ * @param {boolean} button-disabled disables component if it is set to `true`
+ * @param {string} button-style desired style for the main button, i.e. 'che-button-default', 'che-button-primary' etc.
+ *
+ * @example
+ * <example module="userDashboard">
+ *   <file name="index.html">
+ *     <che-button-dropdown button-style="che-button-default"
+ *                          main-action-config="ctrl.mainAction"
+ *                          other-actions-config="ctrl.otherActions"></che-button-dropdown>
+ *   </file>
+ * </example>
+ *
+ * @author Oleksii Kurinnyi
+ */
+export class CheButtonDropdownDirective implements ng.IDirective {
   restrict = 'E';
-  bindToController = true;
-  templateUrl = 'components/widget/button-dropdown/che-button-dropdown.html';
-  controller = 'CheButtonDropdownCtrl';
-  controllerAs = 'cheButtonDropdownCtrl';
 
   // scope values
   scope = {
-    labelText: '@cheButtonDropdownLabel',
-    href: '@cheButtonDropdownHref',
-    ctrl: '=cheButtonDropdownController',
-    isDisabled: '=cheDisabled'
+    mainActionConfig: '=mainActionConfig',
+    otherActionsConfig: '=otherActionsConfig',
+    isDisabled: '=?buttonDisabled',
+    buttonStyle: '@buttonStyle'
   };
+
+  template($element: ng.IAugmentedJQuery,
+           $attrs: ICheButtonDropdownAttributes): string {
+    return `
+<div class="che-button-dropdown btn-group" uib-dropdown>
+
+  <!--type='button'-->
+  <${$attrs.buttonStyle} ng-if="mainActionConfig.type!=='link'"
+                        ng-disabled="isDisabled"
+                        ng-click="mainActionConfig.action()"
+                        class="split-button"
+                        che-button-title="{{mainActionConfig.title}}"
+                        name="{{mainActionConfig.name ? mainActionConfig.name : 'split-button'}}"></${$attrs.buttonStyle}>
+  <!--type='link'-->
+  <${$attrs.buttonStyle} ng-if="mainActionConfig.type==='link'"
+                        ng-disabled="isDisabled"
+                        href="{{mainActionConfig.href}}"
+                        class="split-button"
+                        che-button-title="{{mainActionConfig.title}}"
+                        name="{{mainActionConfig.name ? mainActionConfig.name : 'split-button'}}"></${$attrs.buttonStyle}>
+
+  <${$attrs.buttonStyle} uib-dropdown-toggle
+                        ng-disabled="isDisabled"
+                        id="dropdown-toggle"
+                        class="dropdown"
+                        che-button-title=""
+                        name="dropdown-toggle"
+                        che-button-icon="fa fa-caret-down"></${$attrs.buttonStyle}>
+
+  <ul class="area-dropdown"
+      uib-dropdown-menu
+      role="menu"
+      aria-labelledby="split-button">
+    <li role="menuitem"
+        ng-repeat="otherAction in otherActionsConfig | orderBy:'orderNumber'">
+      <!--type='button'-->
+      <span ng-if="otherAction.type!=='link'"
+            ng-click="otherAction.action()">{{otherAction.title}}</span>
+      <!--type='link'-->
+      <a ng-if="otherAction.type==='link'"
+         href="{{otherAction.href}}">{{otherAction.title}}</a>
+    </li>
+  </ul>
+
+</div>
+    `;
+  }
 
 }
