--- dashboard/src/components/widget/button-dropdown2/che-button-dropdown2.styl
@@ -1,60 +1,0 @@
-.che-button-dropdown2
-  position relative
-  min-width 110px
-
-  button.che-button,
-  .fa-caret-down
-    margin 0 !important
-    min-height 40px
-    float left
-
-  .split-button
-    button.md-accent.md-raised.md-button
-      min-width 72px
-      border-right 1px solid $white-color
-      border-top-right-radius 0 !important
-      border-bottom-right-radius 0 !important
-
-  .dropdown-toggle
-    button
-      padding 0
-      &.md-accent.md-raised.md-button
-        padding 0
-        min-width 38px
-        border-top-left-radius 0 !important
-        border-bottom-left-radius 0 !important
-      md-icon
-        width 100%
-        &:before
-          font-size 15px
-      .fa-caret-down:before
-        position relative
-        top 4px
-    button[disabled] md-icon:before
-      color #9F9F9F !important
-    button:not([disabled]) md-icon:before
-      color $white-color !important
-
-  .area-dropdown
-    background $white-color
-    color $label-secondary-color
-    border-radius 2px
-    left inherit
-    top 42px
-    right 0
-    li
-      outline none
-      cursor pointer
-    li.selected-item
-      background alpha($mouse-gray-color, 0.75)
-      &:hover
-        che-brightness(95%)
-    li:not(.selected-item):hover
-      background alpha($mouse-gray-color, 0.4)
-    span, a
-      display inline-block
-      white-space nowrap
-      line-height 20px
-      padding 0 10px
-      font-size 12px
-      width 100%
